Removed low severity bugs

+++ src/main/java/org/apache/commons/imaging/icc/IccProfileParser.java
@@ -114,7 +114,7 @@ public class IccProfileParser extends BinaryFileParser {
 
             skipBytes(is, 12, "Not a Valid ICC Profile");
 
-            skipBytes(is, 4 * 3);
+            skipBytes(is, (long) 4 * 3);
 
             final int deviceManufacturer = read4Bytes("ProfileFileSignature", is, "Not a Valid ICC Profile", getByteOrder());
             if (LOGGER.isLoggable(Level.FINEST)) {
